Added two packages TFIDFUtils/ and TFIDFClassifier/
TFIDFUtils is just a bunch of utility functions for manipulating
and calculating certain scores for datasets which contain both (1)
a corpus of text, and (2) TFIDF scores for the corpus of text.

The TFIDFClassifier/ package simply trains a Multi-Layer Perceptron
for to classify new HTML documents into either (1) "drug" concept, or
(2) "disease" concept. In the future, this package will be updated to
include several types of classifiers.

The plan:
	- Create weighted-class TFIDF to account for the class imbalance
	  problem
	  OR
	- Simply scrape more disease instances of webpages

	- Solidify the current TFIDF method that we have.

The current TFIDF method that I developed does not show promise because
the prediction scores are too high. Recall, accuracy, and precision, as
well as F1 scores are all around 98%. I will set up a K-Fold Train/Test
split so that we average out all the scores and get a better view of how
the model is performing.